requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain-groq
matplotlib
seaborn